+++ crates/api/src/atom/atom_test.rs
@@ -5,8 +5,8 @@
 use approx::{assert_relative_eq, assert_ulps_eq};
 
 use crate::{
-    term::UNITY_ARRAY, Atom, Classification, Composable, Composition, Dimension, UcumSymbol,
-    UcumUnit,
+    term::UNITY_ARRAY, ucum_unit::UcumUnit, Atom, Classification, Composable, Composition,
+    Dimension, UcumSymbol,
 };
 
 macro_rules! validate_definition {