+++ crates/api/src/term/is_compatible_with.rs
@@ -1,8 +1,8 @@
 use std::borrow::Cow;
 
-use crate::{Composable, IsCompatibleWith, Term};
-
-use super::annotation_composition::AnnotationComposable;
+use crate::{
+    term::annotation_composition::AnnotationComposable, Composable, IsCompatibleWith, Term,
+};
 
 /// In order to enforce compatibility on "non-units" (ex. `{each}`, `{total}`, `{heartbeats}`),
 /// `Term`s need to compare their annotations along with their `Composition`s. In practice, and
@@ -144,3 +144,4 @@ mod tests {
         }
     }
 }
+