+++ crates/atom_generator/src/rust_structs.rs
@@ -1,11 +1,9 @@
-pub(crate) mod pest_symbol_list;
 pub(crate) mod rust_atom_list;
 pub(crate) mod rust_classification_list;
 pub(crate) mod rust_function_set;
 pub(crate) mod rust_property_list;
 
 pub(crate) use self::{
-    pest_symbol_list::PestSymbolList,
     rust_atom_list::{RustAtom, RustAtomList},
     rust_classification_list::RustClassificationList,
     rust_function_set::RustFunctionSet,